Issue #2984453: refactor more drush-8 specific code.
