+++ modules/mongodb/mongodb.drush.inc
@@ -5,4 +5,4 @@
  * A Drush plugin for MongoDB.
  */
 
-require __DIR__ . '/mongodb_drush.php';
+require __DIR__ . '/mongodb.drush.php';
