+++ modules/mongodb/mongodb.drush.yml
@@ -1,0 +1,18 @@
+commands:
+  mongodb_find:
+    aliases: ['mdbf', 'mdbq', 'mongodb:find']
+    arguments:
+      alias: 'The database alias'
+      collection: 'The collection name in the database'
+      selector: 'A MongoDB find() selector in JSON format. Defaults to {}'
+    bootstrap: 4
+    description: 'Execute a query against a collection.'
+    examples:
+      'drush mongodb-find logger watchdog': 'Get the logger/watchdog error-level templates'
+      'drush mongodb-find logger watchdog ''{ "severity": 3 }''': 'Get all the logger/watchdog entries tracking rows.'
+    required_arguments: 2
+
+  mongodb_settings:
+    aliases: ['mdbs', 'mongodb:settings']
+    bootstrap: 3
+    description: 'Print MongoDB settings in Yaml format.'
