+++ modules/mongodb/src/MongoDb.php
@@ -2,6 +2,7 @@
 
 namespace Drupal\mongodb;
 
+use Drupal\mongodb\Commands\MongoDbCommands;
 use MongoDB\Collection;
 
 /**
@@ -20,6 +21,16 @@ class MongoDb {
   const SERVICE_DB_FACTORY = 'mongodb.database_factory';
 
   protected static $libraryVersion;
+
+  /**
+   * Service helper for commands.
+   *
+   * @return \Drupal\mongodb\Commands\MongoDbCommands
+   *   Return the commands service.
+   */
+  public static function commands(): MongoDbCommands {
+    return \Drupal::service(static::SERVICE_COMMANDS);
+  }
 
   /**
    * Guess an approximation of the library version, to handle API changes.
